+++ src/components/BuilderEditor/index.module.css
@@ -10,19 +10,3 @@
 .component:hover {
   border: 1px solid yellow;
 }
-
-.component-border {
-  position: relative;
-}
-
-.type-tag {
-  position: absolute;
-  top: -15px;
-  left: 0;
-  padding: 0.2rem 0.5rem;
-  background-color: yellow;
-  color: black;
-  font-size: 0.7rem;
-
-  font-weight: bold;
-}
