+++ src/components/Grid/index.module.css
@@ -1,4 +1,3 @@
 .grid-container {
     display: grid;
-    gap: 1rem;
 }
