+++ src/components/HOCs/InteractionHandler/index.module.css
@@ -1,0 +1,15 @@
+.component-border {
+    position: relative;
+}
+
+.type-tag {
+    position: absolute;
+    top: -15px;
+    left: 0;
+    padding: 0.2rem 0.5rem;
+    background-color: yellow;
+    color: black;
+    font-size: 0.7rem;
+
+    font-weight: bold;
+}
